Создает дополнительные модули util.js , data.js . Подключает точку входа в index main.js
